requirements.txt
pandas
python-dotenv
python-telegram-bot
python-telegram-bot[job-queue]